cart

+++ src/main/java/com/ravi/orbit/entity/Product.java
@@ -5,6 +5,8 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Entity
 @Getter
 @Setter
@@ -41,16 +43,16 @@ public class Product {
     private String imageUrl;
 
     @Column(name = "market_price")
-    private double marketPrice;
+    private BigDecimal marketPrice;
 
     @Column(name = "discount_percent")
-    private double discountPercent;
+    private BigDecimal discountPercent;
 
     @Column(name = "discount_amount")
-    private double discountAmount;
+    private BigDecimal discountAmount;
 
     @Column(name = "selling_price")
-    private double sellingPrice;
+    private BigDecimal sellingPrice;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", referencedColumnName = "id")
